Change the index type to short

+++ communication-protocol-model/src/main/java/pub/guoxin/protocol/model/anno/CodeIndex.java
@@ -16,7 +16,7 @@ public @interface CodeIndex {
      * <p>
      * 同时该字段也表示该字段在传输时的顺序，顺序从小到大正序排列
      */
-    int index();
+    short index();
 
     /**
      * 描述
